separate view/cache and view/fetch methods
